# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Oasis Garden & Patio API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/inventory:
    get:
      operationId: adminListInventory
      tags: [admin-inventory]
      summary: List products with current inventory levels
      parameters:
        - name: q
          in: query
          required: false
          schema: { type: string }
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [in_stock, low_stock, out_of_stock]
        - name: manufacturerId
          in: query
          required: false
          schema: { type: integer }
        - name: categoryId
          in: query
          required: false
          schema: { type: integer }
        - name: includeInactive
          in: query
          required: false
          schema: { type: boolean, default: false }
        - name: page
          in: query
          required: false
          schema: { type: integer, minimum: 1, default: 1 }
        - name: pageSize
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - name: sortBy
          in: query
          required: false
          schema:
            type: string
            enum: [name, sku, manufacturer, category, onHand, reorderThreshold]
        - name: sortOrder
          in: query
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: asc
      responses:
        "200":
          description: Paginated inventory items
          content: { application/json: { schema: { $ref: "#/components/schemas/AdminInventoryPage" } } }
        "401":
          description: Not authenticated
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "403":
          description: Not admin
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /admin/inventory/adjust:
    post:
      operationId: adminAdjustInventory
      tags: [admin-inventory]
      summary: Atomically apply a manual inventory adjustment + write an audit row
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateInventoryAdjustmentRequest" }
      responses:
        "200":
          description: New inventory level after adjustment
          content: { application/json: { schema: { $ref: "#/components/schemas/AdjustInventoryResponse" } } }
        "400":
          description: Validation error (zero delta, would go negative, unknown location)
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "404":
          description: Product not found
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /admin/inventory/adjustments:
    get:
      operationId: adminListInventoryAdjustments
      tags: [admin-inventory]
      summary: List inventory adjustments (audit trail)
      parameters:
        - name: productId
          in: query
          required: false
          schema: { type: integer }
        - name: locationId
          in: query
          required: false
          schema: { type: integer }
        - name: type
          in: query
          required: false
          schema: { type: string }
        - name: page
          in: query
          required: false
          schema: { type: integer, minimum: 1, default: 1 }
        - name: pageSize
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      responses:
        "200":
          description: Paginated adjustments
          content: { application/json: { schema: { $ref: "#/components/schemas/AdminInventoryAdjustmentsPage" } } }
        "401":
          description: Not authenticated
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /admin/inventory/locations:
    get:
      operationId: adminListInventoryLocations
      tags: [admin-inventory]
      summary: List all inventory locations
      responses:
        "200":
          description: Locations
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/InventoryLocation" }
    post:
      operationId: adminCreateInventoryLocation
      tags: [admin-inventory]
      summary: Create a new inventory location
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateInventoryLocationRequest" }
      responses:
        "201":
          description: Location created
          content: { application/json: { schema: { $ref: "#/components/schemas/InventoryLocation" } } }
        "400":
          description: Validation error
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "409":
          description: Duplicate code
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /admin/inventory/locations/{id}:
    put:
      operationId: adminUpdateInventoryLocation
      tags: [admin-inventory]
      summary: Update a location
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateInventoryLocationRequest" }
      responses:
        "200":
          description: Updated location
          content: { application/json: { schema: { $ref: "#/components/schemas/InventoryLocation" } } }
        "400":
          description: Validation error
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "404":
          description: Not found
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "409":
          description: Duplicate code
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /admin/inventory/locations/{id}/active:
    patch:
      operationId: adminSetInventoryLocationActive
      tags: [admin-inventory]
      summary: Activate or deactivate a location
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SetActiveRequest" }
      responses:
        "200":
          description: Updated location
          content: { application/json: { schema: { $ref: "#/components/schemas/InventoryLocation" } } }
        "400":
          description: Cannot deactivate the default location
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "404":
          description: Not found
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /admin/inventory/locations/{id}/default:
    post:
      operationId: adminSetInventoryLocationDefault
      tags: [admin-inventory]
      summary: Mark a location as the default; clears the flag on all others atomically
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Updated location (now default)
          content: { application/json: { schema: { $ref: "#/components/schemas/InventoryLocation" } } }
        "400":
          description: Location is not active
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "404":
          description: Not found
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    SetActiveRequest:
      type: object
      properties:
        isActive:
          type: boolean
      required:
        - isActive
...
    InventoryLocation:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        code: { type: ["string", "null"] }
        address: { type: ["string", "null"] }
        isActive: { type: boolean }
        isDefault: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, name, code, address, isActive, isDefault, createdAt, updatedAt]
    CreateInventoryLocationRequest:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        code: { type: ["string", "null"] }
        address: { type: ["string", "null"] }
        isActive: { type: boolean, default: true }
      required: [name]
    UpdateInventoryLocationRequest:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        code: { type: ["string", "null"] }
        address: { type: ["string", "null"] }
      required: [name]
    AdminInventoryItem:
      type: object
      description: Product with current stock + computed status.
      properties:
        productId: { type: integer }
        name: { type: string }
        sku: { type: string }
        slug: { type: string }
        manufacturerName: { type: ["string", "null"] }
        categoryName: { type: ["string", "null"] }
        primaryImageUrl: { type: ["string", "null"] }
        onHand: { type: integer }
        onHold: { type: integer }
        lowStockThreshold: { type: integer }
        reorderThreshold: { type: integer }
        status:
          type: string
          enum: [in_stock, low_stock, out_of_stock]
        isActive: { type: boolean }
        updatedAt: { type: ["string", "null"], format: date-time }
      required:
        - productId
        - name
        - sku
        - slug
        - manufacturerName
        - categoryName
        - primaryImageUrl
        - onHand
        - onHold
        - lowStockThreshold
        - reorderThreshold
        - status
        - isActive
        - updatedAt
    AdminInventoryPage:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/AdminInventoryItem" }
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
      required: [items, total, page, pageSize]
    InventoryAdjustmentType:
      type: string
      description: Allowed adjustment categories for manual changes.
      enum:
        - cycle_count
        - damage
        - loss
        - found
        - transfer
        - return
        - manual_correction
        - other
    CreateInventoryAdjustmentRequest:
      type: object
      properties:
        productId: { type: integer }
        locationId:
          type: ["integer", "null"]
          description: Defaults to current default location when null.
        adjustmentType: { $ref: "#/components/schemas/InventoryAdjustmentType" }
        quantityChange:
          type: integer
          description: Signed delta (positive to add, negative to remove). Cannot be zero.
        reason: { type: ["string", "null"] }
      required: [productId, adjustmentType, quantityChange]
    AdjustInventoryResponse:
      type: object
      properties:
        productId: { type: integer }
        onHand: { type: integer }
        adjustmentId: { type: integer }
      required: [productId, onHand, adjustmentId]
...
    AdminInventoryAdjustmentsPage:
      type: object
      properties:
        adjustments:
          type: array
          items: { $ref: "#/components/schemas/AdminInventoryAdjustment" }
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
      required: [adjustments, total, page, pageSize]
...